requirements.txt
uvicorn==0.23.2
pydantic==2.9.2
langchain-core
langchain_google_genai
langgraph==0.2.46
python-dotenv==1.0.1
assistant-stream==0.0.5
fastapi==0.115.11
qdrant-client==1.13.3
langchain-community==0.3.20
python-multipart==0.0.20
langchain-text-splitters==0.3.7
pypdf==5.4.0
docx2txt==0.8
unstructured==0.17.2
pymongo==4.7.0
langchain-qdrant